Feature prepare test os

+++ docker/create_mymid.sh
@@ -1,5 +1,12 @@
 #!/bin/sh
 
+echo "check docker"
+if ! command -v docker >/dev/null 2>&1; then
+    echo "docker not exists,please install"
+    echo "For TLinxu : yum install docker-ce"
+    exit 100
+fi
+
 echo "clear image"
 docker stop mymid
 docker rm mymid
